op-deployer,ctb: Migrate DeployDisputeGame to new script API
